backend/requirements.txt
# Django 框架
Django>=4.2,<5.0
djangorestframework>=3.14

# MongoDB 支持
mongoengine>=0.27
pymongo>=4.6

# CORS 支持
django-cors-headers>=4.3

# 日期时间处理
python-dateutil>=2.8

# 数据处理
pandas>=2.0
openpyxl>=3.1

# 生产服务器
gunicorn>=21.0

# 开发工具
ipython>=8.0